// src/constants/enums/routesPath.ts
enum RoutesPath {
  root = '/',
  registration = '/registration',
  login = '/login',
  privacyPolicy = '/privacy-policy',
  termsOfUse = '/terms-of-use',
  map = '/map',
  chat = '/chat',
  friends = '/friends',
  accessGeo = '/access-geo',
  accessGeoError = '/access-geo-error',
  profile = '/profile',
  accessAge = '/access-age',
  comingSoon = '/coming-soon',
}

export default RoutesPath;